Refactor: atom 파일 구조
