+++ Lotte-Cinema/src/components/timeSelect/TimeInfo.tsx
@@ -1,6 +1,5 @@
-import CinemaTypeBar from './atom/CinemaTypeBar';
-import MovieTimeBox from './atom/MovieTimeBox';
-import TimeTypeBar from './atom/TimeTypeBar';
+import { CinemaTypeBar, TimeTypeBar } from './atom/bar';
+import MovieTimeBox from './atom/infoBox/MovieTimeBox';
 
 // "상영관 타입바 / 상영시간 타입바 / 상영관별 시간 박스"를 담은 컴포넌트입니다.
 const TimeInfo = () => {
