+++ Lotte-Cinema/src/components/timeSelect/atom/bar/CinemaTypeBar.tsx
