+++ Lotte-Cinema/src/components/timeSelect/atom/bar/TimeTypeBar.tsx
