+++ Lotte-Cinema/src/components/timeSelect/atom/bar/index.ts
@@ -1,0 +1,4 @@
+import CinemaTypeBar from './CinemaTypeBar';
+import TimeTypeBar from './TimeTypeBar';
+
+export { CinemaTypeBar, TimeTypeBar };
