+++ Lotte-Cinema/src/components/timeSelect/atom/infoBox/MovieTimeBox.tsx
