+++ Lotte-Cinema/src/components/timeSelect/atom/infoBox/TimeAccordion.tsx
