+++ Lotte-Cinema/src/components/timeSelect/atom/infoBox/TimeCard.tsx
