+++ Lotte-Cinema/src/pages/TimeSelect/TimeSelect.tsx
@@ -1,8 +1,6 @@
 import Header from '@/components/commons/header/Header';
 import MobileLayout from '@/components/mobileLayout/MobileLayout';
-import Calendar from '@/components/timeSelect/Calendar';
-import MovieInfoBar from '@/components/timeSelect/MovieInfoBar';
-import TimeInfo from '@/components/timeSelect/TimeInfo';
+import { Calendar, MovieInfoBar, TimeInfo } from '@/components/timeSelect';
 
 const TimeSelect = () => {
   return (
